renamed 'workouts' route to 'workout-list'
